refactor: CQRS 패턴 적용
