+++ src/main/java/com/phcworld/phcworldboardservice/controller/response/FreeBoardResponse.java
@@ -1,9 +1,9 @@
-package com.phcworld.phcworldboardservice.controller.port;
+package com.phcworld.phcworldboardservice.controller.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.phcworld.phcworldboardservice.domain.FreeBoard;
-import com.phcworld.phcworldboardservice.service.port.FreeBoardAnswerResponse;
-import com.phcworld.phcworldboardservice.service.port.UserResponse;
+import com.phcworld.phcworldboardservice.service.dto.FreeBoardAnswerResponse;
+import com.phcworld.phcworldboardservice.service.dto.UserResponse;
 import com.phcworld.phcworldboardservice.utils.LocalDateTimeUtils;
 import lombok.Builder;
