+++ src/main/java/com/phcworld/phcworldboardservice/messagequeue/dto/Schema.java
@@ -1,4 +1,4 @@
-package com.phcworld.phcworldboardservice.messagequeue.port;
+package com.phcworld.phcworldboardservice.messagequeue.dto;
 
 import lombok.Builder;
 import lombok.Data;
