+++ src/main/java/com/phcworld/phcworldboardservice/service/dto/FreeBoardAnswerResponse.java
@@ -1,6 +1,5 @@
-package com.phcworld.phcworldboardservice.service.port;
+package com.phcworld.phcworldboardservice.service.dto;
 
-import com.phcworld.phcworldboardservice.service.port.UserResponse;
 import lombok.Builder;
 
 @Builder
