+++ src/main/java/com/phcworld/phcworldboardservice/service/port/FreeBoardRepository.java
@@ -1,8 +1,7 @@
 package com.phcworld.phcworldboardservice.service.port;
 
-import com.phcworld.phcworldboardservice.controller.port.FreeBoardSearch;
+import com.phcworld.phcworldboardservice.infrastructure.dto.FreeBoardSearch;
 import com.phcworld.phcworldboardservice.domain.FreeBoard;
-import com.phcworld.phcworldboardservice.domain.User;
 import org.springframework.data.domain.Pageable;
 
 import java.util.List;
