--- src/main/java/com/phcworld/phcworldboardservice/service/port/SystemDateTimeHolder.java
@@ -1,13 +1,0 @@
-package com.phcworld.phcworldboardservice.service.port;
-
-import org.springframework.stereotype.Component;
-
-import java.time.LocalDateTime;
-
-@Component
-public class SystemDateTimeHolder implements LocalDateTimeHolder{
-    @Override
-    public LocalDateTime now() {
-        return LocalDateTime.now();
-    }
-}
