+++ src/test/java/com/phcworld/phcworldboardservice/domain/port/FreeBoardRequestTest.java
@@ -1,5 +1,6 @@
 package com.phcworld.phcworldboardservice.domain.port;
 
+import com.phcworld.phcworldboardservice.domain.FreeBoardRequest;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
@@ -11,7 +12,6 @@ import org.junit.jupiter.api.Test;
 import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class FreeBoardRequestTest {
