+++ src/test/java/com/phcworld/phcworldboardservice/mock/FakeFreeBoardRepository.java
@@ -1,6 +1,6 @@
 package com.phcworld.phcworldboardservice.mock;
 
-import com.phcworld.phcworldboardservice.controller.port.FreeBoardSearch;
+import com.phcworld.phcworldboardservice.infrastructure.dto.FreeBoardSearch;
 import com.phcworld.phcworldboardservice.domain.FreeBoard;
 import com.phcworld.phcworldboardservice.service.port.FreeBoardRepository;
 import org.springframework.data.domain.Pageable;
