+++ src/test/java/com/phcworld/phcworldboardservice/mock/TestContainer.java
@@ -1,12 +1,12 @@
 package com.phcworld.phcworldboardservice.mock;
 
-import com.phcworld.phcworldboardservice.controller.FreeBoardApiController;
+import com.phcworld.phcworldboardservice.controller.FreeBoardCommandApiController;
+import com.phcworld.phcworldboardservice.controller.FreeBoardQueryApiController;
 import com.phcworld.phcworldboardservice.controller.port.FreeBoardService;
 import com.phcworld.phcworldboardservice.controller.port.WebclientService;
 import com.phcworld.phcworldboardservice.service.FreeBoardServiceImpl;
 import com.phcworld.phcworldboardservice.service.port.FreeBoardRepository;
 import com.phcworld.phcworldboardservice.service.port.LocalDateTimeHolder;
-import com.phcworld.phcworldboardservice.service.port.UserRepository;
 import lombok.Builder;
 
 public class TestContainer {
@@ -15,9 +15,11 @@ public class TestContainer {
 
     public final FreeBoardService freeBoardService;
 
-    public final FreeBoardApiController freeBoardApiController;
+    public final FreeBoardCommandApiController freeBoardCommandApiController;
 
     public final WebclientService webclientService;
+
+    public final FreeBoardQueryApiController freeBoardQueryApiController;
 
     @Builder
     public TestContainer(LocalDateTimeHolder localDateTimeHolder){
@@ -28,7 +30,11 @@ public class TestContainer {
                 .localDateTimeHolder(localDateTimeHolder)
                 .boardProducer(new FakeKafkaProducer())
                 .build();
-        this.freeBoardApiController = FreeBoardApiController.builder()
+        this.freeBoardQueryApiController = FreeBoardQueryApiController.builder()
+                .webclientService(webclientService)
+                .freeBoardService(freeBoardService)
+                .build();
+        this.freeBoardCommandApiController = FreeBoardCommandApiController.builder()
                 .webclientService(webclientService)
                 .freeBoardService(freeBoardService)
                 .build();
